requirements.txt
# Core dependencies for Phase 1A NFL Ingestion

# Data processing
pandas>=2.0.0
pyarrow>=12.0.0  # For parquet file support

# NFL data sources
nfl-data-py>=0.3.0  # nflverse Python wrapper

# Configuration
pyyaml>=6.0

# Machine Learning
scikit-learn>=1.0.0
xgboost>=1.5.0  # Optional, falls back to sklearn if not available

# Testing
pytest>=7.0.0

